+++ includes/views/print/tmpl-receipt.php
@@ -166,7 +166,7 @@
         {{#with meta}}
         <dl class="meta">
           {{#each []}}
-          <dt>{{label}}:</dt>
+          <dt>{{#if label}}{{label}}{{else}}{{key}}{{/if}}:</dt>
           <dd>{{value}}</dd>
           {{/each}}
         </dl>