+++ changelog.md
@@ -1,3 +1,8 @@
+= 0.4.19 = - 2018/09/24
+* Fix: Gutenberg conflict causing error on POS settings page
+* Fix: remove customer meta for POS
+* Fix: tweak payment options on Order edit page
+
 = 0.4.18 = - 2018/02/12
 * Fix: Product variation stock adjustment
 * Fix: Category search for products